Reflex.Collection

Widgets on Collections

listHoldWithKey :: forall t m k v a. (Ord k, Adjustable t m, MonadHold t m) => Map k v -> Event t (Map k (Maybe v)) -> (k -> v -> m a) -> m (Dynamic t (Map k a))

listWithKey :: forall t k v m a. (Ord k, Adjustable t m, PostBuild t m, MonadFix m, MonadHold t m) => Dynamic t (Map k v) -> (k -> Dynamic t v -> m a) -> m (Dynamic t (Map k a))

listWithKey' :: (Ord k, Adjustable t m, MonadFix m, MonadHold t m) => Map k v -> Event t (Map k (Maybe v)) -> (k -> v -> Event t v -> m a) -> m (Dynamic t (Map k a))

Deprecated: listWithKey' has been renamed to listWithKeyShallowDiff; also, its behavior has changed to fix a bug where children were always rebuilt (never updated)

listWithKeyShallowDiff :: (Ord k, Adjustable t m, MonadFix m, MonadHold t m) => Map k v -> Event t (Map k (Maybe v)) -> (k -> v -> Event t v -> m a) -> m (Dynamic t (Map k a))

Display the given map of items (in key order) using the builder function provided, and update it with the given event. Nothing update entries will delete the corresponding children, and Just entries will create them if they do not exist or send an update event to them if they do.

listViewWithKey :: (Ord k, Adjustable t m, PostBuild t m, MonadHold t m, MonadFix m) => Dynamic t (Map k v) -> (k -> Dynamic t v -> m (Event t a)) -> m (Event t (Map k a))

Create a dynamically-changing set of Event-valued widgets. This is like listWithKey, specialized for widgets returning (Event t a). listWithKey would return 'Dynamic t (Map k (Event t a))' in this scenario, but listViewWithKey flattens this to 'Event t (Map k a)' via switch.

selectViewListWithKey

Arguments

:: (Adjustable t m, Ord k, PostBuild t m, MonadHold t m, MonadFix m)

=> Dynamic t k
Current selection key

-> Dynamic t (Map k v)
Dynamic key/value map

-> (k -> Dynamic t v -> Dynamic t Bool -> m (Event t a))
Function to create a widget for a given key from Dynamic value and Dynamic Bool indicating if this widget is currently selected

-> m (Event t (k, a))
Event that fires when any child's return Event fires. Contains key of an arbitrary firing widget.

Create a dynamically-changing set of widgets, one of which is selected at any time.

selectViewListWithKey_

Arguments

:: (Adjustable t m, Ord k, PostBuild t m, MonadHold t m, MonadFix m)

=> Dynamic t k
Current selection key

-> Dynamic t (Map k v)
Dynamic key/value map

-> (k -> Dynamic t v -> Dynamic t Bool -> m (Event t a))
Function to create a widget for a given key from Dynamic value and Dynamic Bool indicating if this widget is currently selected

-> m (Event t k)
Event that fires when any child's return Event fires. Contains key of an arbitrary firing widget.

List Utils

list :: (Ord k, Adjustable t m, MonadHold t m, PostBuild t m, MonadFix m) => Dynamic t (Map k v) -> (Dynamic t v -> m a) -> m (Dynamic t (Map k a))

Create a dynamically-changing set of widgets from a Dynamic key/value map. Unlike the withKey variants, the child widgets are insensitive to which key they're associated with.

simpleList :: (Adjustable t m, MonadHold t m, PostBuild t m, MonadFix m) => Dynamic t [v] -> (Dynamic t v -> m a) -> m (Dynamic t [a])

Create a dynamically-changing set of widgets from a Dynamic list.