Reflex.DynamicWriter

Description

This module defines MonadDynamicWriter and DynamicWriterT, its standard implementation.

Documentation

class (Monad m, Monoid w) => MonadDynamicWriter t w m | m -> t w where

MonadDynamicWriter efficiently collects Dynamic values using tellDyn and combines them monoidally to provide a Dynamic result.

Minimal complete definition

tellDyn

Methods

tellDyn :: Dynamic t w -> m ()

newtype DynamicWriterT t w m a

A basic implementation of MonadDynamicWriter.

Constructors

DynamicWriterT

Fields

unDynamicWriterT :: StateT [Dynamic t w] m a

runDynamicWriterT :: (MonadFix m, Reflex t, Monoid w) => DynamicWriterT t w m a -> m (a, Dynamic t w)

Run a DynamicWriterT action. The dynamic writer output will be provided along with the result of the action.

withDynamicWriterT :: (Monoid w, Monoid w', Reflex t, MonadHold t m, MonadFix m) => (w -> w') -> DynamicWriterT t w m a -> DynamicWriterT t w' m a

Map a function over the output of a DynamicWriterT.