Reflex.FunctorMaybe

Description

This module defines the FunctorMaybe class, which extends Functors with the ability to delete values.

Documentation

class FunctorMaybe f where

A class for values that combines filtering and mapping using Maybe. Morally, FunctorMaybe ~ KleisliFunctor Maybe. Also similar is the Witherable typeclass, but it requires Foldable f and Traverable f, and e.g. Event is instance of neither.

A definition of fmapMaybe must satisfy the following laws:

identity

fmapMaybe Just ≡ id

composition

fmapMaybe (f <=< g) ≡ fmapMaybe f . fmapMaybe g

Minimal complete definition

fmapMaybe

Methods

fmapMaybe :: (a -> Maybe b) -> f a -> f b

Combined mapping and filtering function.