Reflex.Network

Documentation

networkView :: (Reflex t, NotReady t m, Adjustable t m, PostBuild t m) => Dynamic t (m a) -> m (Event t a)

Given a Dynamic of network-creating actions, create a network that is recreated whenever the Dynamic updates. The returned Event of network results occurs when the Dynamic does. Note: Often, the type a is an Event, in which case the return value is an Event-of-Events that would typically be flattened (via switchPromptly).

networkHold :: (Reflex t, Adjustable t m, MonadHold t m) => m a -> Event t (m a) -> m (Dynamic t a)

Given an initial network and an Event of network-creating actions, create a network that is recreated whenever the Event fires. The returned Dynamic of network results occurs when the Event does. Note: Often, the type a is an Event, in which case the return value is a Dynamic-of-Events that would typically be flattened.

untilReady :: (Reflex t, Adjustable t m, PostBuild t m) => m a -> m b -> m (a, Event t b)

Render a placeholder network to be shown while another network is not yet done building