Reflex.Patch

Description

This module defines the Patch class, which is used by Reflex to manage changes to Incremental values.

Documentation

class (Semigroup q, Monoid q) => Group q where

A Group is a Monoid where every element has an inverse.

Minimal complete definition

negateG

Methods

negateG :: q -> q

(~~) :: q -> q -> q

class Semigroup q => Additive q

An Additive Semigroup is one where (<>) is commutative

newtype AdditivePatch p

The elements of an Additive Semigroup can be considered as patches of their own type.

Constructors

AdditivePatch

Fields

unAdditivePatch :: p