Reflex.Patch.Class

Description

The interface for types which represent changes made to other types

Documentation

class Patch p where

A Patch type represents a kind of change made to a datastructure.

Minimal complete definition

apply

Associated Types

type PatchTarget p :: *

Methods

apply :: p -> PatchTarget p -> Maybe (PatchTarget p)

Apply the patch p a to the value a. If no change is needed, return Nothing.

applyAlways :: Patch p => p -> PatchTarget p -> PatchTarget p

Apply a Patch; if it does nothing, return the original value