Reflex.Patch.DMap

Documentation

newtype PatchDMap k v

A set of changes to a DMap. Any element may be inserted/updated or deleted.

Constructors

PatchDMap

Fields

unPatchDMap :: DMap k (ComposeMaybe v)

mapPatchDMap :: (forall a. v a -> v' a) -> PatchDMap k v -> PatchDMap k v'

traversePatchDMap :: Applicative m => (forall a. v a -> m (v' a)) -> PatchDMap k v -> m (PatchDMap k v')

traversePatchDMapWithKey :: Applicative m => (forall a. k a -> v a -> m (v' a)) -> PatchDMap k v -> m (PatchDMap k v')

weakenPatchDMapWith :: (forall a. v a -> v') -> PatchDMap k v -> PatchMap (Some k) v'

patchDMapToPatchMapWith :: (f v -> v') -> PatchDMap (Const2 k v) f -> PatchMap k v'

const2PatchDMapWith :: forall k v f a. (v -> f a) -> PatchMap k v -> PatchDMap (Const2 k a) f

const2IntPatchDMapWith :: forall v f a. (v -> f a) -> PatchIntMap v -> PatchDMap (Const2 Key a) f

getDeletions :: GCompare k => PatchDMap k v -> DMap k v' -> DMap k v'

Get the values that will be deleted if the given patch is applied to the given DMap. Includes values that will be replaced.