Reflex.Patch.DMapWithMove

Documentation

newtype PatchDMapWithMove k v

Like PatchMapWithMove, but for DMap.

Constructors

PatchDMapWithMove (DMap k (NodeInfo k v))

data NodeInfo k v a

Constructors

NodeInfo

Fields

_nodeInfo_from :: !(From k v a)

_nodeInfo_to :: !(To k a)

data From k v :: a -> * where

Constructors

From_Insert :: v a -> From k v a

From_Delete :: From k v a

From_Move :: !(k a) -> From k v a

type To = ComposeMaybe

validPatchDMapWithMove :: forall k v. (GCompare k, EqTag k (ComposeMaybe k)) => PatchDMapWithMove k v -> Bool

data Pair1 f g a

Constructors

Pair1 (f a) (g a)

data Fixup k v a

Constructors

Fixup_Delete

Fixup_Update (These (From k v a) (To k a))

getHere :: These a b -> Maybe a

getThere :: These a b -> Maybe b

moveDMapKey :: GCompare k => k a -> k a -> PatchDMapWithMove k v

deleteDMapKey :: k a -> PatchDMapWithMove k v

unPatchDMapWithMove :: PatchDMapWithMove k v -> DMap k (NodeInfo k v)

unsafePatchDMapWithMove :: DMap k (NodeInfo k v) -> PatchDMapWithMove k v

Warning: when using this function, you must ensure that the invariants of PatchDMapWithMove are preserved; they will not be checked.

mapPatchDMapWithMove :: forall k v v'. (forall a. v a -> v' a) -> PatchDMapWithMove k v -> PatchDMapWithMove k v'

traversePatchDMapWithMove :: forall m k v v'. Applicative m => (forall a. v a -> m (v' a)) -> PatchDMapWithMove k v -> m (PatchDMapWithMove k v')

traversePatchDMapWithMoveWithKey :: forall m k v v'. Applicative m => (forall a. k a -> v a -> m (v' a)) -> PatchDMapWithMove k v -> m (PatchDMapWithMove k v')

nodeInfoMapFrom :: (From k v a -> From k v' a) -> NodeInfo k v a -> NodeInfo k v' a

nodeInfoMapFromM :: Functor f => (From k v a -> f (From k v' a)) -> NodeInfo k v a -> f (NodeInfo k v' a)

weakenPatchDMapWithMoveWith :: forall k v v'. (forall a. v a -> v') -> PatchDMapWithMove k v -> PatchMapWithMove (Some k) v'

patchDMapWithMoveToPatchMapWithMoveWith :: forall k f v v'. (f v -> v') -> PatchDMapWithMove (Const2 k v) f -> PatchMapWithMove k v'

const2PatchDMapWithMoveWith :: forall k v f a. (v -> f a) -> PatchMapWithMove k v -> PatchDMapWithMove (Const2 k a) f

getDeletionsAndMoves :: GCompare k => PatchDMapWithMove k v -> DMap k v' -> DMap k (Product v' (ComposeMaybe k))

Get the values that will be deleted or moved if the given patch is applied to the given DMap.