Reflex.Patch.Map

Description

Patches on Map that consist only of insertions (including overwrites) and deletions

Documentation

newtype PatchMap k v

A set of changes to a Map. Any element may be inserted/updated or deleted. Insertions are represented as values wrapped in Just, while deletions are represented as Nothings

Constructors

PatchMap

Fields

unPatchMap :: Map k (Maybe v)

patchMapNewElements :: PatchMap k v -> [v]

Returns all the new elements that will be added to the Map

patchMapNewElementsMap :: PatchMap k v -> Map k v

Returns all the new elements that will be added to the Map