Reflex.Patch.MapWithMove

Description

Patches on Map that can insert, delete, and move values from one key to another

Documentation

newtype PatchMapWithMove k v

Patch a DMap with additions, deletions, and moves. Invariant: If key k1 is coming from From_Move k2, then key k2 should be going to Just k1, and vice versa. There should never be any unpaired From/To keys.

Constructors

PatchMapWithMove (Map k (NodeInfo k v))

data NodeInfo k v

Holds the information about each key: where its new value should come from, and where its old value should go to

Constructors

NodeInfo

Fields

_nodeInfo_from :: !(From k v)

Where do we get the new value for this key?

_nodeInfo_to :: !(To k)

If the old value is being kept (i.e. moved rather than deleted or replaced), where is it going?

patchMapWithMove :: Ord k => Map k (NodeInfo k v) -> Maybe (PatchMapWithMove k v)

Create a PatchMapWithMove, validating it

unPatchMapWithMove :: PatchMapWithMove k v -> Map k (NodeInfo k v)

Extract the internal representation of the PatchMapWithMove

unsafePatchMapWithMove :: Map k (NodeInfo k v) -> PatchMapWithMove k v

Warning: when using this function, you must ensure that the invariants of PatchMapWithMove are preserved; they will not be checked.

data From k v

Describe how a key's new value should be produced

Constructors

From_Insert v
Insert the given value here

From_Delete
Delete the existing value, if any, from here

From_Move !k
Move the value here from the given key

type To = Maybe

Describe where a key's old value will go. If this is Just, that means the key's old value will be moved to the given other key; if it is Nothing, that means it will be deleted.

patchMapWithMoveNewElements :: PatchMapWithMove k v -> [v]

Returns all the new elements that will be added to the Map

patchMapWithMoveNewElementsMap :: PatchMapWithMove k v -> Map k v

patchThatSortsMapWith :: Ord k => (v -> v -> Ordering) -> Map k v -> PatchMapWithMove k v

Create a PatchMapWithMove that, if applied to the given Map, will sort its values using the given ordering function. The set keys of the Map is not changed.

patchThatChangesAndSortsMapWith :: forall k v. (Ord k, Ord v) => (v -> v -> Ordering) -> Map k v -> Map k v -> PatchMapWithMove k v

Create a PatchMapWithMove that, if applied to the first Map provided, will produce a Map with the same values as the second Map but with the values sorted with the given ordering function.

nodeInfoMapFrom :: (From k v -> From k v) -> NodeInfo k v -> NodeInfo k v

Change the From value of a NodeInfo

nodeInfoMapMFrom :: Functor f => (From k v -> f (From k v)) -> NodeInfo k v -> f (NodeInfo k v)

Change the From value of a NodeInfo, using a Functor (or Applicative, Monad, etc.) action to get the new value

nodeInfoSetTo :: To k -> NodeInfo k v -> NodeInfo k v

Set the To field of a NodeInfo