Reflex.PostBuild.Class

Description

This module defines PostBuild, which indicates that an action will be notified when it, and any action it's a part of, has finished executing.

Documentation

class (Reflex t, Monad m) => PostBuild t m | m -> t where

PostBuild represents an action that is notified via an Event when it has finished executing. Note that the specific definition of "finished" is determined by the instance of PostBuild, but the intent is to allow Behaviors and Dynamics to be safely sampled, regardless of where they were created, when the post-build Event fires. The post-build Event will fire exactly once for an given action.

Minimal complete definition

getPostBuild

Methods

getPostBuild :: m (Event t ())

Retrieve the post-build Event for this action.