Reflex.Pure

Description

This module provides a pure implementation of Reflex, which is intended to serve as a reference for the semantics of the Reflex class. All implementations of Reflex should produce the same results as this implementation, although performance and laziness/strictness may differ.

Documentation

data Pure t

A completely pure-functional Reflex timeline, identifying moments in time with the type t.

Orphan instances

MonadSample (Pure t) ((->) t)

Methods

sample :: Behavior (Pure t) a -> t -> a