Reflex.Query.Class

Documentation

class (Monoid (QueryResult a), Semigroup (QueryResult a)) => Query a where

Minimal complete definition

crop

Associated Types

type QueryResult a :: *

Methods

crop :: a -> QueryResult a -> QueryResult a

data QueryMorphism q q'

NB: QueryMorphism's must be group homomorphisms when acting on the query type and compatible with the query relationship when acting on the query result

Constructors

QueryMorphism

Fields

_queryMorphism_mapQuery :: q -> q'

_queryMorphism_mapQueryResult :: QueryResult q' -> QueryResult q

newtype SelectedCount

This type keeps track of the multiplicity of elements of the view selector that are being used by the app

Constructors

SelectedCount

Fields

unSelectedCount :: Int

combineSelectedCounts :: SelectedCount -> SelectedCount -> Maybe SelectedCount

The SemigroupMonoidGroup instances for a ViewSelector should use this function which returns Nothing if the result is 0. This allows the pruning of leaves that are no longer wanted.

class (Group q, Additive q, Query q) => MonadQuery t q m | m -> q t where

Minimal complete definition

tellQueryIncremental, askQueryResult, queryIncremental

Methods

tellQueryIncremental :: Incremental t (AdditivePatch q) -> m ()

askQueryResult :: m (Dynamic t (QueryResult q))

queryIncremental :: Incremental t (AdditivePatch q) -> m (Dynamic t (QueryResult q))

tellQueryDyn :: (Reflex t, MonadQuery t q m) => Dynamic t q -> m ()

queryDyn :: (Reflex t, Monad m, MonadQuery t q m) => Dynamic t q -> m (Dynamic t (QueryResult q))