Reflex.Spider

Description

This module exports all of the user-facing functionality of the Spider Reflex engine

Documentation

type Spider = SpiderTimeline Global

The default, global Spider environment

data SpiderTimeline x

Designates the default, global Spider timeline

data Global

A statically allocated SpiderTimeline; this is the only instance of SpiderTimeline that can used when using -fspecialize-to-spidertimeline-global.

data SpiderHost x a

The monad for actions that manipulate a Spider timeline identified by x

runSpiderHost :: SpiderHost Global a -> IO a

Run an action affecting the global Spider timeline; this will be guarded by a mutex for that timeline

runSpiderHostForTimeline :: SpiderHost x a -> SpiderTimelineEnv x -> IO a

Run an action affecting a given Spider timeline; this will be guarded by a mutex for that timeline

newSpiderTimeline :: IO (Some SpiderTimelineEnv)

Create a new SpiderTimelineEnv

withSpiderTimeline :: (forall x. HasSpiderTimeline x => SpiderTimelineEnv x -> IO r) -> IO r

Pass a new timeline to the given function.

Deprecated

type SpiderEnv = SpiderTimeline

Deprecated: Use SpiderTimelineEnv instead

SpiderEnv is the old name for SpiderTimeline