Reflex.TriggerEvent.Class

Description

This module defines TriggerEvent, which describes actions that may create new Events that can be triggered from IO.

Documentation

class Monad m => TriggerEvent t m | m -> t where

TriggerEvent represents actions that can create Events that can be triggered by IO actions.

Minimal complete definition

newTriggerEvent, newTriggerEventWithOnComplete, newEventWithLazyTriggerWithOnComplete

Methods

newTriggerEvent :: m (Event t a, a -> IO ())

Create a triggerable Event. Whenever the resulting function is called, the resulting Event will fire at some point in the future. Note that this may not be synchronous.

newTriggerEventWithOnComplete :: m (Event t a, a -> IO () -> IO ())

Like newTriggerEvent, but the callback itself takes another callback, to be invoked once the requested Event occurrence has finished firing. This allows synchronous operation.

newEventWithLazyTriggerWithOnComplete :: ((a -> IO () -> IO ()) -> IO (IO ())) -> m (Event t a)

Like newTriggerEventWithOnComplete, but with setup and teardown. This relatively complex type signature allows any external listeners to be subscribed lazily and then removed whenever the returned Event is no longer being listened to. Note that the setup/teardown may happen multiple times, and there is no guarantee that the teardown will be executed promptly, or even at all, in the case of program termination.