Data.Functor.Misc

Description

This module provides types and functions with no particular theme, but which are relevant to the use of Functor-based datastructures like DMap.

Const2

data Const2 :: * -> x -> x -> * where

Const2 stores a value of a given type k and ensures that a particular type v is always given for the last type parameter

Constructors

Const2 :: k -> Const2 k v v

unConst2 :: Const2 k v v' -> k

dmapToMap :: DMap (Const2 k v) Identity -> Map k v

Convert a DMap to a regular Map

dmapToMapWith :: (f v -> v') -> DMap (Const2 k v) f -> Map k v'

mapToDMap :: Map k v -> DMap (Const2 k v) Identity

Convert a regular Map to a DMap

weakenDMapWith :: (forall a. v a -> v') -> DMap k v -> Map (Some k) v'

WrapArg

data WrapArg :: (k -> *) -> (k -> *) -> * -> * where

WrapArg can be used to tag a value in one functor with a type representing another functor. This was primarily used with dependent-map < 0.2, in which the value type was not wrapped in a separate functor.

Constructors

WrapArg :: f a -> WrapArg g f (g a)

Convenience functions for DMap

mapWithFunctorToDMap :: Map k (f v) -> DMap (Const2 k v) f

Convert a regular Map, where the values are already wrapped in a functor, to a DMap

mapKeyValuePairsMonotonic :: (DSum k v -> DSum k' v') -> DMap k v -> DMap k' v'

Map over all key/value pairs in a DMap, potentially altering the key as well as the value. The provided function MUST preserve the ordering of the keys, or the resulting DMap will be malformed.

combineDMapsWithKey :: forall f g h i. GCompare f => (forall a. f a -> These (g a) (h a) -> i a) -> DMap f g -> DMap f h -> DMap f i

Union two DMaps of different types, yielding another type. Each key that is present in either input map will be present in the output.

data EitherTag l r a where

Tag type for Either to use it as a DSum.

Constructors

LeftTag :: EitherTag l r l

RightTag :: EitherTag l r r

dmapToThese :: DMap (EitherTag a b) Identity -> Maybe (These a b)

Extract the values of a DMap of EitherTags.

eitherToDSum :: Either a b -> DSum (EitherTag a b) Identity

Convert Either to a DSum. Inverse of dsumToEither.

dsumToEither :: DSum (EitherTag a b) Identity -> Either a b

Convert DSum to Either. Inverse of eitherToDSum.

Deprecated functions

sequenceDmap :: Applicative t => DMap f t -> t (DMap f Identity)

Deprecated: Use 'Data.Dependent.Map.traverseWithKey (_ -> fmap Identity)' instead

Run the actions contained in the DMap

wrapDMap :: (forall a. a -> f a) -> DMap k Identity -> DMap k f

Deprecated: Use 'Data.Dependent.Map.map (f . runIdentity)' instead

Replace the Identity functor for a DMap's values with a different functor

rewrapDMap :: (forall a. f a -> g a) -> DMap k f -> DMap k g

Deprecated: Use map instead

Replace one functor for a DMap's values with a different functor

unwrapDMap :: (forall a. f a -> a) -> DMap k f -> DMap k Identity

Deprecated: Use 'Data.Dependent.Map.map (Identity . f)' instead

Replace one functor for a DMap's values with the Identity functor

unwrapDMapMaybe :: GCompare k => (forall a. f a -> Maybe a) -> DMap k f -> DMap k Identity

Deprecated: Use 'Data.Dependent.Map.mapMaybeWithKey (_ a -> fmap Identity $ f a)' instead

Like unwrapDMap, but possibly delete some values from the DMap

extractFunctorDMap :: DMap (Const2 k (f v)) Identity -> DMap (Const2 k v) f

Deprecated: Use 'mapKeyValuePairsMonotonic ((Const2 k :=> Identity v) -> Const2 k :=> v)' instead

Eliminate the Identity functor in a DMap and replace it with the underlying functor

newtype ComposeMaybe f a

We can't use Compose Maybe instead of ComposeMaybe, because that would make the f parameter have a nominal type role. We need f to be representational so that we can use safe coerce.

Constructors

ComposeMaybe

Fields

getComposeMaybe :: Maybe (f a)