Reflex.Class

Description

This module contains the Reflex interface, as well as a variety of convenience functions for working with Events, Behaviors, and other signals.

Documentation

module Reflex.Patch

Primitives

class (MonadHold t (PushM t), MonadSample t (PullM t), MonadFix (PushM t), Functor (Dynamic t), Applicative (Dynamic t), Monad (Dynamic t)) => Reflex t where

The Reflex class contains all the primitive functionality needed for Functional Reactive Programming (FRP). The t type parameter indicates which "timeline" is in use. Timelines are fully-independent FRP contexts, and the type of the timeline determines the FRP engine to be used. For most purposes, the Spider implementation is recommended.

Minimal complete definition

never, constant, push, pushCheap, pull, merge, fan, switch, coincidence, current, updated, unsafeBuildDynamic, unsafeBuildIncremental, mergeIncremental, mergeIncrementalWithMove, currentIncremental, updatedIncremental, incrementalToDynamic, behaviorCoercion, eventCoercion, dynamicCoercion

Associated Types

data Behavior t :: * -> *

A container for a value that can change over time. Behaviors can be sampled at will, but it is not possible to be notified when they change

data Event t :: * -> *

A stream of occurrences. During any given frame, an Event is either occurring or not occurring; if it is occurring, it will contain a value of the given type (its "occurrence type")

data Dynamic t :: * -> *

A container for a value that can change over time and allows notifications on changes. Basically a combination of a Behavior and an Event, with a rule that the Behavior will change if and only if the Event fires.

data Incremental t :: * -> *

An Incremental is a more general form of a Dynamic. Instead of always fully replacing the value, only parts of it can be patched. This is only needed for performance critical code via mergeIncremental to make small changes to large values.

type PushM t :: * -> *

A monad for doing complex push-based calculations efficiently

type PullM t :: * -> *

A monad for doing complex pull-based calculations efficiently

Methods

never :: Event t a

An Event with no occurrences

constant :: a -> Behavior t a

Create a Behavior that always has the given value

push :: (a -> PushM t (Maybe b)) -> Event t a -> Event t b

Create an Event from another Event; the provided function can sample Behaviors and hold Events, and use the results to produce a occurring (Just) or non-occurring (Nothing) result

pushCheap :: (a -> PushM t (Maybe b)) -> Event t a -> Event t b

Like push but intended for functions that the implementation can consider cheap to compute for performance considerations. WARNING: The function passed to pushCheap may be run multiple times without any caching.

pull :: PullM t a -> Behavior t a

Create a Behavior by reading from other Behaviors; the result will be recomputed whenever any of the read Behaviors changes

merge :: GCompare k => DMap k (Event t) -> Event t (DMap k Identity)

Merge a collection of events; the resulting Event will only occur if at least one input event is occuring, and will contain all of the input keys that are occurring simultaneously

fan :: GCompare k => Event t (DMap k Identity) -> EventSelector t k

Efficiently fan-out an event to many destinations. This function should be partially applied, and then the result applied repeatedly to create child events

switch :: Behavior t (Event t a) -> Event t a

Create an Event that will occur whenever the currently-selected input Event occurs

coincidence :: Event t (Event t a) -> Event t a

Create an Event that will occur whenever the input event is occurring and its occurrence value, another Event, is also occurring

current :: Dynamic t a -> Behavior t a

Extract the Behavior of a Dynamic.

updated :: Dynamic t a -> Event t a

Extract the Event of the Dynamic.

unsafeBuildDynamic :: PullM t a -> Event t a -> Dynamic t a

Create a new Dynamic. The given PullM must always return the most recent firing of the given Event, if any.

unsafeBuildIncremental :: Patch p => PullM t (PatchTarget p) -> Event t p -> Incremental t p

Create a new Incremental. The given PullM's value must always change in the same way that the accumulated application of patches would change that value.

mergeIncremental :: GCompare k => Incremental t (PatchDMap k (Event t)) -> Event t (DMap k Identity)

Create a merge whose parents can change over time

mergeIncrementalWithMove :: GCompare k => Incremental t (PatchDMapWithMove k (Event t)) -> Event t (DMap k Identity)

Experimental: Create a merge whose parents can change over time; changing the key of an Event is more efficient than with mergeIncremental

currentIncremental :: Patch p => Incremental t p -> Behavior t (PatchTarget p)

Extract the Behavior component of an Incremental

updatedIncremental :: Patch p => Incremental t p -> Event t p

Extract the Event component of an Incremental

incrementalToDynamic :: Patch p => Incremental t p -> Dynamic t (PatchTarget p)

Convert an Incremental to a Dynamic

behaviorCoercion :: Coercion a b -> Coercion (Behavior t a) (Behavior t b)

Construct a Coercion for a Behavior given an Coercion for its occurrence type

eventCoercion :: Coercion a b -> Coercion (Event t a) (Event t b)

Construct a Coercion for an Event given an Coercion for its occurrence type

dynamicCoercion :: Coercion a b -> Coercion (Dynamic t a) (Dynamic t b)

Construct a Coercion for a Dynamic given an Coercion for its occurrence type

coerceBehavior :: (Reflex t, Coercible a b) => Behavior t a -> Behavior t b

Coerce a Behavior between representationally-equivalent value types

coerceEvent :: (Reflex t, Coercible a b) => Event t a -> Event t b

Coerce an Event between representationally-equivalent occurrence types

coerceDynamic :: (Reflex t, Coercible a b) => Dynamic t a -> Dynamic t b

Coerce a Dynamic between representationally-equivalent value types

class (Applicative m, Monad m) => MonadSample t m | m -> t where

MonadSample designates monads that can read the current value of a Behavior. This includes both PullM and PushM.

Minimal complete definition

sample

Methods

sample :: Behavior t a -> m a

Get the current value in the Behavior

class MonadSample t m => MonadHold t m where

MonadHold designates monads that can create new Behaviors based on Events; usually this will be PushM or a monad based on it. MonadHold is required to create any stateful computations with Reflex.

Methods

hold :: a -> Event t a -> m (Behavior t a)

Create a new Behavior whose value will initially be equal to the given value and will be updated whenever the given Event occurs. The update takes effect immediately after the Event occurs; if the occurrence that sets the Behavior (or one that is simultaneous with it) is used to sample the Behavior, it will see the _old_ value of the Behavior, not the new one.

hold :: (m ~ f m', MonadTrans f, MonadHold t m') => a -> Event t a -> m (Behavior t a)

Create a new Behavior whose value will initially be equal to the given value and will be updated whenever the given Event occurs. The update takes effect immediately after the Event occurs; if the occurrence that sets the Behavior (or one that is simultaneous with it) is used to sample the Behavior, it will see the _old_ value of the Behavior, not the new one.

holdDyn :: a -> Event t a -> m (Dynamic t a)

Create a Dynamic value using the given initial value that changes every time the Event occurs.

holdDyn :: (m ~ f m', MonadTrans f, MonadHold t m') => a -> Event t a -> m (Dynamic t a)

Create a Dynamic value using the given initial value that changes every time the Event occurs.

holdIncremental :: Patch p => PatchTarget p -> Event t p -> m (Incremental t p)

Create an Incremental value using the given initial value that changes every time the Event occurs.

holdIncremental :: (Patch p, m ~ f m', MonadTrans f, MonadHold t m') => PatchTarget p -> Event t p -> m (Incremental t p)

Create an Incremental value using the given initial value that changes every time the Event occurs.

buildDynamic :: PullM t a -> Event t a -> m (Dynamic t a)

buildDynamic :: (m ~ f m', MonadTrans f, MonadHold t m') => PullM t a -> Event t a -> m (Dynamic t a)

fan-related types

newtype EventSelector t k

An EventSelector allows you to efficiently select an Event based on a key. This is much more efficient than filtering for each key with fmapMaybe.

Constructors

EventSelector

Fields

select :: forall a. k a -> Event t a

Retrieve the Event for the given key. The type of the Event is determined by the type of the key, so this can be used to fan-out Events whose sub-Events have different types.

Using EventSelectors and the fan primitive is far more efficient than (but equivalent to) using fmapMaybe to select only the relevant occurrences of an Event.

Incremental-related types

Convenience functions

constDyn :: Reflex t => a -> Dynamic t a

Construct a Dynamic value that never changes

pushAlways :: Reflex t => (a -> PushM t b) -> Event t a -> Event t b

Create an Event from another Event. The provided function can sample Behaviors and hold Events.

Combining Events

leftmost :: Reflex t => [Event t a] -> Event t a

Create a new Event that occurs if at least one of the Events in the list occurs. If multiple occur at the same time the value is the value of the leftmost event.

mergeMap :: (Reflex t, Ord k) => Map k (Event t a) -> Event t (Map k a)

Create a new Event combining the map of Events into an Event that occurs if at least one of them occurs and has a map of values of all Events occuring at that time.

mergeMapIncremental :: (Reflex t, Ord k) => Incremental t (PatchMap k (Event t a)) -> Event t (Map k a)

Create a merge whose parents can change over time

mergeMapIncrementalWithMove :: (Reflex t, Ord k) => Incremental t (PatchMapWithMove k (Event t a)) -> Event t (Map k a)

Experimental: Create a merge whose parents can change over time; changing the key of an Event is more efficient than with mergeIncremental

mergeList :: Reflex t => [Event t a] -> Event t (NonEmpty a)

Create a new Event that occurs if at least one of the Events in the list occurs and has a list of the values of all Events occuring at that time.

mergeWith :: Reflex t => (a -> a -> a) -> [Event t a] -> Event t a

Create a new Event that occurs if at least one of the Events in the list occurs. If multiple occur at the same time they are folded from the left with the given function.

difference :: Reflex t => Event t a -> Event t b -> Event t a

Create a new Event that occurs when the first supplied Event occurs unless the second supplied Event occurs simultaneously.

alignEventWithMaybe :: Reflex t => (These a b -> Maybe c) -> Event t a -> Event t b -> Event t c

Breaking up Events

splitE :: Reflex t => Event t (a, b) -> (Event t a, Event t b)

Split the supplied Event into two individual Events occuring at the same time with the respective values from the tuple.

fanEither :: Reflex t => Event t (Either a b) -> (Event t a, Event t b)

Split the event into separate events for Left and Right values.

fanThese :: Reflex t => Event t (These a b) -> (Event t a, Event t b)

Split the event into separate events for This and That values, allowing them to fire simultaneously when the input value is These.

fanMap :: (Reflex t, Ord k) => Event t (Map k a) -> EventSelector t (Const2 k a)

Split the event into an EventSelector that allows efficient selection of the individual Events.

dmapToThese :: DMap (EitherTag a b) Identity -> Maybe (These a b)

Extract the values of a DMap of EitherTags.

data EitherTag l r a where

Tag type for Either to use it as a DSum.

Constructors

LeftTag :: EitherTag l r l

RightTag :: EitherTag l r r

eitherToDSum :: Either a b -> DSum (EitherTag a b) Identity

Convert Either to a DSum. Inverse of dsumToEither.

dsumToEither :: DSum (EitherTag a b) Identity -> Either a b

Convert DSum to Either. Inverse of eitherToDSum.

Switching between Events

switchPromptly :: (Reflex t, MonadHold t m) => Event t a -> Event t (Event t a) -> m (Event t a)

Switches to the new event whenever it receives one. Whenever a new event is provided, if it is firing, its value will be the resulting event's value; if it is not firing, but the old one is, the old one's value will be used.

switchPromptOnly :: (Reflex t, MonadHold t m) => Event t a -> Event t (Event t a) -> m (Event t a)

switches to a new event whenever it receives one. At the moment of switching, the old event will be ignored if it fires, and the new one will be used if it fires; this is the opposite of switch, which will use only the old value.

Using Events to sample Behaviors

tag :: Reflex t => Behavior t b -> Event t a -> Event t b

Replace each occurrence value of the Event with the value of the Behavior at the time of that occurrence.

attach :: Reflex t => Behavior t a -> Event t b -> Event t (a, b)

Create a new Event that combines occurences of supplied Event with the current value of the Behavior.

attachWith :: Reflex t => (a -> b -> c) -> Behavior t a -> Event t b -> Event t c

Create a new Event that occurs when the supplied Event occurs by combining it with the current value of the Behavior.

attachWithMaybe :: Reflex t => (a -> b -> Maybe c) -> Behavior t a -> Event t b -> Event t c

Create a new Event by combining each occurence with the current value of the Behavior. The occurrence is discarded if the combining function returns Nothing

Blocking an Event based on a Behavior

gate :: Reflex t => Behavior t Bool -> Event t a -> Event t a

Create a new Event that only occurs if the supplied Event occurs and the Behavior is true at the time of occurence.

Combining Dynamics

distributeDMapOverDynPure :: forall t k. (Reflex t, GCompare k) => DMap k (Dynamic t) -> Dynamic t (DMap k Identity)

This function converts a DMap whose elements are Dynamics into a Dynamic DMap. Its implementation is more efficient than doing the same through the use of multiple uses of zipWithDyn or Applicative operators.

distributeListOverDyn :: Reflex t => [Dynamic t a] -> Dynamic t [a]

Convert a list of Dynamics into a Dynamic list.

distributeListOverDynWith :: Reflex t => ([a] -> b) -> [Dynamic t a] -> Dynamic t b

Create a new Dynamic by applying a combining function to a list of Dynamics

zipDyn :: Reflex t => Dynamic t a -> Dynamic t b -> Dynamic t (a, b)

Combine two Dynamics. The result will change whenever either (or both) input Dynamic changes. Equivalent to zipDynWith (,).

zipDynWith :: Reflex t => (a -> b -> c) -> Dynamic t a -> Dynamic t b -> Dynamic t c

Combine two Dynamics with a combining function. The result will change whenever either (or both) input Dynamic changes. More efficient than liftA2.

Accumulating state

class Reflex t => Accumulator t f | f -> t where

An Accumulator type can be built by accumulating occurrences of an Event.

Minimal complete definition

accumMaybeM, mapAccumMaybeM

Methods

accum :: (MonadHold t m, MonadFix m) => (a -> b -> a) -> a -> Event t b -> m (f a)

accumM :: (MonadHold t m, MonadFix m) => (a -> b -> PushM t a) -> a -> Event t b -> m (f a)

accumMaybe :: (MonadHold t m, MonadFix m) => (a -> b -> Maybe a) -> a -> Event t b -> m (f a)

accumMaybeM :: (MonadHold t m, MonadFix m) => (a -> b -> PushM t (Maybe a)) -> a -> Event t b -> m (f a)

mapAccum :: (MonadHold t m, MonadFix m) => (a -> b -> (a, c)) -> a -> Event t b -> m (f a, Event t c)

mapAccumM :: (MonadHold t m, MonadFix m) => (a -> b -> PushM t (a, c)) -> a -> Event t b -> m (f a, Event t c)

mapAccumMaybe :: (MonadHold t m, MonadFix m) => (a -> b -> (Maybe a, Maybe c)) -> a -> Event t b -> m (f a, Event t c)

mapAccumMaybeM :: (MonadHold t m, MonadFix m) => (a -> b -> PushM t (Maybe a, Maybe c)) -> a -> Event t b -> m (f a, Event t c)

mapAccum_ :: forall t m a b c. (Reflex t, MonadHold t m, MonadFix m) => (a -> b -> (a, c)) -> a -> Event t b -> m (Event t c)

Accumulate occurrences of an Event, producing an output occurrence each time. Discard the underlying Accumulator.

mapAccumM_ :: forall t m a b c. (Reflex t, MonadHold t m, MonadFix m) => (a -> b -> PushM t (a, c)) -> a -> Event t b -> m (Event t c)

Accumulate occurrences of an Event, using a PushM action and producing an output occurrence each time. Discard the underlying Accumulator.

mapAccumMaybe_ :: forall t m a b c. (Reflex t, MonadHold t m, MonadFix m) => (a -> b -> (Maybe a, Maybe c)) -> a -> Event t b -> m (Event t c)

Accumulate occurrences of an Event, possibly producing an output occurrence each time. Discard the underlying Accumulator.

mapAccumMaybeM_ :: forall t m a b c. (Reflex t, MonadHold t m, MonadFix m) => (a -> b -> PushM t (Maybe a, Maybe c)) -> a -> Event t b -> m (Event t c)

Accumulate occurrences of an Event, using a PushM action and possibly producing an output occurrence each time. Discard the underlying Accumulator.

zipListWithEvent :: (Reflex t, MonadHold t m, MonadFix m) => (a -> b -> c) -> [a] -> Event t b -> m (Event t c)

Create a new Event by combining each occurence with the next value of the list using the supplied function. If the list runs out of items, all subsequent Event occurrences will be ignored.

numberOccurrences :: (Reflex t, MonadHold t m, MonadFix m, Num b) => Event t a -> m (Event t (b, a))

Assign a number to each occurence of the given Event, starting from 0

numberOccurrencesFrom :: (Reflex t, MonadHold t m, MonadFix m, Num b) => b -> Event t a -> m (Event t (b, a))

Assign a number to each occurence of the given Event

(<@>) :: Reflex t => Behavior t (a -> b) -> Event t a -> Event t b infixl 4

This is used to sample the value of a Behavior using an Event.

The <@> operator is intended to be used in conjunction with the Applicative instance for Behavior.

This is useful when we want to combine the values of one Event and the value of several Behaviors at the time the Event is firing.

If we have:

f  :: a -> b -> c -> d
b1 :: Behavior t a
b2 :: Behavior t b
e  :: Event t c

then we can do:

f <$> b1 <*> b2 <@> e :: Event t d

in order to apply the function f to the relevant values.

The alternative would be something like:

attachWith (\(x1, x2) y -> f x1 x2 y) ((,) <$> b1 <*> b2) e :: Event t d

or a variation involing a custom data type to hold the combination of Behaviors even when that combination might only ever be used by f.

A more suggestive example might be:

handleMouse <$> bPlayerState <*> bMousePosition <@> eMouseClick :: Event t (GameState -> GameState)

(<@) :: Reflex t => Behavior t b -> Event t a -> Event t b infixl 4

An version of <@> that does not use the value of the Event.

Alternatively, it is tag in operator form.

This is useful when we want to combine the values of several Behaviors at particular points in time using an Applicative style syntax.

If we have:

g  :: a -> b -> d
b1 :: Behavior t a
b2 :: Behavior t b
e  :: Event t c

where e is firing at the points in time of interest.

Then we can use <@:

g <$> b1 <*> b2 <@  e :: Event t d

to combine the values of b1 and b2 at each of those points of time, with the function g being used to combine the values.

This is the same as <@> except that the Event is being used only to act as a trigger.

headE :: (Reflex t, MonadHold t m, MonadFix m) => Event t a -> m (Event t a)

Create a new Event that only occurs only once, on the first occurence of the supplied Event.

tailE :: (Reflex t, MonadHold t m, MonadFix m) => Event t a -> m (Event t a)

Create a new Event that occurs on all but the first occurence of the supplied Event.

headTailE :: (Reflex t, MonadHold t m, MonadFix m) => Event t a -> m (Event t a, Event t a)

Create a tuple of two Events with the first one occuring only the first time the supplied Event occurs and the second occuring on all but the first occurence.

takeWhileE :: forall t m a. (Reflex t, MonadFix m, MonadHold t m) => (a -> Bool) -> Event t a -> m (Event t a)

Starting at the current time, fire all the occurrences of the Event for which the given predicate returns True. When False is returned, do not fire, and permanently stop firing, even if True values are encountered later.

switcher :: (Reflex t, MonadHold t m) => Behavior t a -> Event t (Behavior t a) -> m (Behavior t a)

Create a new behavior given a starting behavior and switch to a the behvior carried by the event when it fires.

Debugging functions

traceEvent :: (Reflex t, Show a) => String -> Event t a -> Event t a

Print the supplied String and the value of the Event on each occurence. This should only be used for debugging.

Note: As with Debug.Trace.trace, the message will only be printed if the Event is actually used.

traceEventWith :: Reflex t => (a -> String) -> Event t a -> Event t a

Print the output of the supplied function on each occurence of the Event. This should only be used for debugging.

Note: As with Debug.Trace.trace, the message will only be printed if the Event is actually used.

Unsafe functions

unsafeDynamic :: Reflex t => Behavior t a -> Event t a -> Dynamic t a

Construct a Dynamic from a Behavior and an Event. The Behavior _must_ change when and only when the Event fires, such that the Behavior's value is always equal to the most recent firing of the Event; if this is not the case, the resulting Dynamic will behave nondeterministically.

unsafeMapIncremental :: (Reflex t, Patch p, Patch p') => (PatchTarget p -> PatchTarget p') -> (p -> p') -> Incremental t p -> Incremental t p'

FunctorMaybe

class FunctorMaybe f where

A class for values that combines filtering and mapping using Maybe. Morally, FunctorMaybe ~ KleisliFunctor Maybe. Also similar is the Witherable typeclass, but it requires Foldable f and Traverable f, and e.g. Event is instance of neither.

A definition of fmapMaybe must satisfy the following laws:

identity

fmapMaybe Just ≡ id

composition

fmapMaybe (f <=< g) ≡ fmapMaybe f . fmapMaybe g

Minimal complete definition

fmapMaybe

Methods

fmapMaybe :: (a -> Maybe b) -> f a -> f b

Combined mapping and filtering function.

fforMaybe :: FunctorMaybe f => f a -> (a -> Maybe b) -> f b

Flipped version of fmapMaybe.

ffilter :: FunctorMaybe f => (a -> Bool) -> f a -> f a

Filter 'f a' using the provided predicate. Relies on fforMaybe.

Miscellaneous convenience functions

ffor :: Functor f => f a -> (a -> b) -> f b

Flipped version of fmap.

class (Reflex t, Monad m) => MonadAdjust t m | m -> t where

A Monad that supports adjustment over time. After an action has been run, if the given events fire, it will adjust itself so that its net effect is as though it had originally been run with the new value. Note that there is some issue here with persistent side-effects: obviously, IO (and some other side-effects) cannot be undone, so it is up to the instance implementer to determine what the best meaning for this class is in such cases.

Minimal complete definition

runWithReplace, traverseDMapWithKeyWithAdjust, traverseDMapWithKeyWithAdjustWithMove

Methods

runWithReplace :: m a -> Event t (m b) -> m (a, Event t b)

traverseDMapWithKeyWithAdjust :: GCompare k => (forall a. k a -> v a -> m (v' a)) -> DMap k v -> Event t (PatchDMap k v) -> m (DMap k v', Event t (PatchDMap k v'))

traverseDMapWithKeyWithAdjustWithMove :: GCompare k => (forall a. k a -> v a -> m (v' a)) -> DMap k v -> Event t (PatchDMapWithMove k v) -> m (DMap k v', Event t (PatchDMapWithMove k v'))

sequenceDMapWithAdjust :: (GCompare k, MonadAdjust t m) => DMap k m -> Event t (PatchDMap k m) -> m (DMap k Identity, Event t (PatchDMap k Identity))

sequenceDMapWithAdjustWithMove :: (GCompare k, MonadAdjust t m) => DMap k m -> Event t (PatchDMapWithMove k m) -> m (DMap k Identity, Event t (PatchDMapWithMove k Identity))

mapMapWithAdjustWithMove :: forall t m k v v'. (MonadAdjust t m, Ord k) => (k -> v -> m v') -> Map k v -> Event t (PatchMapWithMove k v) -> m (Map k v', Event t (PatchMapWithMove k v'))

Deprecated functions

appendEvents :: (Reflex t, Monoid a) => Event t a -> Event t a -> Event t a

Deprecated: If a 'Semigroup a' instance is available, use mappend; otherwise, use 'alignWith (mergeThese mappend)' instead

Create a new Event that occurs if at least one of the supplied Events occurs. If both occur at the same time they are combined using mappend.

onceE :: (Reflex t, MonadHold t m, MonadFix m) => Event t a -> m (Event t a)

Deprecated: Use headE instead

Alias for headE

sequenceThese :: Monad m => These (m a) (m b) -> m (These a b)

Deprecated: Use bisequenceA or bisequence from the bifunctors package instead

Run both sides of a These monadically, combining the results.

fmapMaybeCheap :: Reflex t => (a -> Maybe b) -> Event t a -> Event t b

fmapCheap :: Reflex t => (a -> b) -> Event t a -> Event t b

fforCheap :: Reflex t => Event t a -> (a -> b) -> Event t b

fforMaybeCheap :: Reflex t => Event t a -> (a -> Maybe b) -> Event t b

pushAlwaysCheap :: Reflex t => (a -> PushM t b) -> Event t a -> Event t b

tagCheap :: Reflex t => Behavior t b -> Event t a -> Event t b

mergeWithCheap :: Reflex t => (a -> a -> a) -> [Event t a] -> Event t a

mergeWithCheap' :: Reflex t => (a -> b) -> (b -> b -> b) -> [Event t a] -> Event t b