Reflex.Patch.MapWithMove

Documentation

newtype PatchMapWithMove k v

Patch a DMap with additions, deletions, and moves. Invariants: * The same source key may not be moved to two destination keys. * Any key that is the source of a Move must also be edited (otherwise, the Move would duplicate it)

Constructors

PatchMapWithMove (Map k (NodeInfo k v))

data NodeInfo k v

Constructors

NodeInfo

Fields

_nodeInfo_from :: !(From k v)

_nodeInfo_to :: !(To k)

unPatchMapWithMove :: PatchMapWithMove k v -> Map k (NodeInfo k v)

unsafePatchMapWithMove :: Map k (NodeInfo k v) -> PatchMapWithMove k v

Warning: when using this function, you must ensure that the invariants of PatchMapWithMove are preserved; they will not be checked.

data From k v

Constructors

From_Insert v
Insert the given value here

From_Delete
Delete the existing value, if any, from here

From_Move !k
Move the value here from the given key

type To = Maybe

patchMapWithMoveNewElements :: PatchMapWithMove k v -> [v]

Returns all the new elements that will be added to the Map

validPatchMapWithMove :: Ord k => PatchMapWithMove k v -> Bool

knownValidPatchMapWithMove :: [PatchMapWithMove Int Int]

knownInvalidPatchMapWithMove :: [PatchMapWithMove Int Int]

patchThatSortsMapWith :: Ord k => (v -> v -> Ordering) -> Map k v -> PatchMapWithMove k v

patchThatChangesAndSortsMapWith :: forall k v. (Ord k, Ord v) => (v -> v -> Ordering) -> Map k v -> Map k v -> PatchMapWithMove k v

nodeInfoMapFrom :: (From k v -> From k v) -> NodeInfo k v -> NodeInfo k v

nodeInfoMapMFrom :: Functor f => (From k v -> f (From k v)) -> NodeInfo k v -> f (NodeInfo k v)

nodeInfoSetTo :: To k -> NodeInfo k v -> NodeInfo k v