Reflex.PerformEvent.Base

Description

This module provides PerformEventT, the standard implementation of PerformEvent.

Documentation

newtype PerformEventT t m a

Provides a basic implementation of PerformEvent. Note that, despite the name, PerformEventT is not an instance of MonadTrans.

Constructors

PerformEventT

Fields

unPerformEventT :: RequesterT t (HostFrame t) Identity (HostFrame t) a

newtype FireCommand t m

A function that fires events for the given EventTriggers and then runs any followup actions provided via PerformEvent. The given ReadPhase action will be run once for the initial trigger execution as well as once for each followup.

Constructors

FireCommand

Fields

runFireCommand :: forall a. [DSum (EventTrigger t) Identity] -> ReadPhase m a -> m [a]

hostPerformEventT :: forall t m a. (Monad m, MonadSubscribeEvent t m, MonadReflexHost t m, MonadRef m, Ref m ~ Ref IO) => PerformEventT t m a -> m (a, FireCommand t m)

Run a PerformEventT action, returning a FireCommand that allows the caller to trigger Events while ensuring that performEvent actions are run at the appropriate time.