Reflex.Requester.Base

Description

This module provides RequesterT, the standard implementation of Requester.

Documentation

newtype RequesterT t request response m a

A basic implementation of Requester.

Constructors

RequesterT

Fields

unRequesterT :: EventWriterT t (DMap (Tag (PrimState m)) request) (ReaderT (EventSelector t (WrapArg response (Tag (PrimState m)))) m) a

runRequesterT :: (Reflex t, Monad m) => RequesterT t request response m a -> Event t (DMap (Tag (PrimState m)) response) -> m (a, Event t (DMap (Tag (PrimState m)) request))

Run a RequesterT action. The resulting Event will fire whenever requests are made, and responses should be provided in the input Event. The Tag keys will be used to return the responses to the same place the requests were issued.

runWithReplaceRequesterTWith :: forall m t request response a b. (Reflex t, MonadHold t m) => (forall a' b'. m a' -> Event t (m b') -> RequesterT t request response m (a', Event t b')) -> RequesterT t request response m a -> Event t (RequesterT t request response m b) -> RequesterT t request response m (a, Event t b)

sequenceDMapWithAdjustRequesterTWith :: forall k t request response m v v' p p'. (GCompare k, Reflex t, MonadHold t m, PatchTarget (p' (Some k) (Event t (DMap (Tag (PrimState m)) request))) ~ Map (Some k) (Event t (DMap (Tag (PrimState m)) request)), Patch (p' (Some k) (Event t (DMap (Tag (PrimState m)) request)))) => (forall k' v1 v2. GCompare k' => (forall a. k' a -> v1 a -> m (v2 a)) -> DMap k' v1 -> Event t (p k' v1) -> RequesterT t request response m (DMap k' v2, Event t (p k' v2))) -> (forall v1. (forall a. v1 a -> v' a) -> p k v1 -> p k v') -> (forall v1 v2. (forall a. v1 a -> v2) -> p k v1 -> p' (Some k) v2) -> (forall v2. p' (Some k) v2 -> [v2]) -> (forall a. Incremental t (p' (Some k) (Event t a)) -> Event t (Map (Some k) a)) -> (forall a. k a -> v a -> RequesterT t request response m (v' a)) -> DMap k v -> Event t (p k v) -> RequesterT t request response m (DMap k v', Event t (p k v'))