Reflex.Spider.Internal

Description

This module is the implementation of the Spider Reflex engine. It uses a graph traversal algorithm to propagate Events and Behaviors.

Documentation

debugPropagate :: Bool

debugInvalidateHeight :: Bool

debugInvalidate :: Bool

showNodeId :: a -> String

data EventSubscription x

Constructors

EventSubscription

Fields

_eventSubscription_unsubscribe :: !(IO ())

_eventSubscription_subscribed :: !(EventSubscribed x)

unsubscribe :: EventSubscription x -> IO ()

newtype Event x a

Constructors

Event

Fields

unEvent :: Subscriber x a -> EventM x (EventSubscription x, Maybe a)

subscribeAndRead :: Event x a -> Subscriber x a -> EventM x (EventSubscription x, Maybe a)

pushCheap :: HasSpiderTimeline x => (a -> ComputeM x (Maybe b)) -> Event x a -> Event x b

Construct an Event equivalent to that constructed by push, but with no caching; if the computation function is very cheap, this is (much) more efficient than push

data CacheSubscribed x a

Constructors

CacheSubscribed

Fields

_cacheSubscribed_subscribers :: !(WeakBag (Subscriber x a))

_cacheSubscribed_weakSelf :: !(IORef (Weak (CacheSubscribed x a)))

_cacheSubscribed_parent :: !(EventSubscription x)

_cacheSubscribed_cached :: !(IORef (Maybe (CacheSubscribed x a)))

_cacheSubscribed_occurrence :: !(IORef (Maybe a))

cacheEvent :: HasSpiderTimeline x => Event x a -> Event x a

Construct an Event whose value is guaranteed not to be recomputed repeatedly

TODO: Try a caching strategy where we subscribe directly to the parent when there's only one subscriber, and then build our own WeakBag only when a second subscriber joins

cleanupCacheSubscribed :: CacheSubscribed x a -> IO ()

subscribe :: Event x a -> Subscriber x a -> EventM x (EventSubscription x)

wrap :: MonadIO m => (t -> EventSubscribed x) -> (Subscriber x a -> m (WeakBagTicket, t, Maybe a)) -> Subscriber x a -> m (EventSubscription x, Maybe a)

eventRoot :: GCompare k => k a -> Root x k -> Event x a

eventNever :: Event x a

eventFan :: (GCompare k, HasSpiderTimeline x) => k a -> Fan x k -> Event x a

eventSwitch :: HasSpiderTimeline x => Switch x a -> Event x a

eventCoincidence :: HasSpiderTimeline x => Coincidence x a -> Event x a

eventHold :: HasSpiderTimeline x => Hold x p -> Event x p

eventDyn :: (HasSpiderTimeline x, Patch p) => Dyn x p -> Event x p

subscribeCoincidenceInner :: CanSubscribe x m => Event x a -> Height -> CoincidenceSubscribed x a -> m (Maybe a, Height, EventSubscribed x)

data Subscriber x a

Constructors

Subscriber

Fields

subscriberPropagate :: !(a -> EventM x ())

subscriberInvalidateHeight :: !(Height -> IO ())

subscriberRecalculateHeight :: !(Height -> IO ())

newSubscriberHold :: (HasSpiderTimeline x, Patch p) => Hold x p -> IO (Subscriber x p)

newSubscriberFan :: forall x k. (HasSpiderTimeline x, GCompare k) => FanSubscribed x k -> IO (Subscriber x (DMap k Identity))

newSubscriberSwitch :: forall x a. HasSpiderTimeline x => SwitchSubscribed x a -> IO (Subscriber x a)

newSubscriberCoincidenceOuter :: forall x b. HasSpiderTimeline x => CoincidenceSubscribed x b -> IO (Subscriber x (Event x b))

newSubscriberCoincidenceInner :: forall x a. HasSpiderTimeline x => CoincidenceSubscribed x a -> IO (Subscriber x a)

invalidateSubscriberHeight :: Height -> Subscriber x a -> IO ()

recalculateSubscriberHeight :: Height -> Subscriber x a -> IO ()

propagate :: a -> WeakBag (Subscriber x a) -> EventM x ()

Propagate everything at the current height

toAny :: a -> Any

data EventSubscribed x

Constructors

EventSubscribed

Fields

eventSubscribedHeightRef :: !(IORef Height)

eventSubscribedRetained :: !Any

eventSubscribedRoot :: RootSubscribed x a -> EventSubscribed x

eventSubscribedNever :: EventSubscribed x

eventSubscribedFan :: FanSubscribed x k -> EventSubscribed x

eventSubscribedSwitch :: SwitchSubscribed x a -> EventSubscribed x

eventSubscribedCoincidence :: CoincidenceSubscribed x a -> EventSubscribed x

getEventSubscribedHeight :: EventSubscribed x -> IO Height

subscribeHoldEvent :: CanSubscribe x m => Hold x p -> Subscriber x p -> m (EventSubscription x, Maybe p)

newtype Behavior x a

Constructors

Behavior

Fields

readBehaviorTracked :: BehaviorM x a

behaviorHold :: Hold x p -> Behavior x (PatchTarget p)

behaviorHoldIdentity :: Hold x (Identity a) -> Behavior x a

behaviorConst :: a -> Behavior x a

behaviorPull :: Pull x a -> Behavior x a

behaviorDyn :: Patch p => Dyn x p -> Behavior x (PatchTarget p)

readHoldTracked :: Hold x p -> BehaviorM x (PatchTarget p)

readBehaviorUntracked :: Defer (SomeHoldInit x) m => Behavior x a -> m a

data Dynamic x p

Constructors

Dynamic

Fields

dynamicCurrent :: !(Behavior x (PatchTarget p))

dynamicUpdated :: !(Event x p)

dynamicHold :: HasSpiderTimeline x => Hold x p -> Dynamic x p

dynamicHoldIdentity :: HasSpiderTimeline x => Hold x (Identity a) -> Dynamic x (Identity a)

dynamicConst :: PatchTarget p -> Dynamic x p

dynamicDyn :: (HasSpiderTimeline x, Patch p) => Dyn x p -> Dynamic x p

dynamicDynIdentity :: HasSpiderTimeline x => Dyn x (Identity a) -> Dynamic x (Identity a)

data Hold x p

Constructors

Hold

Fields

holdValue :: !(IORef (PatchTarget p))

holdInvalidators :: !(IORef [Weak (Invalidator x)])

holdEvent :: Event x p

holdParent :: !(IORef (Maybe (EventSubscription x)))

data Global

A statically allocated SpiderTimeline; this is the only instance of SpiderTimeline that can used when using -fspecialize-to-spidertimeline-global.

globalSpiderTimelineEnv :: SpiderTimelineEnv Global

data SpiderTimelineEnv x

Stores all global data relevant to a particular Spider timeline; only one value should exist for each type x

Constructors

SpiderTimelineEnv

Fields

_spiderTimeline_lock :: !(MVar ())

_spiderTimeline_eventEnv :: !(EventEnv x)

data EventEnv x

Constructors

EventEnv

Fields

eventEnvAssignments :: !(IORef [SomeAssignment x])

eventEnvHoldInits :: !(IORef [SomeHoldInit x])

eventEnvDynInits :: !(IORef [SomeDynInit x])

eventEnvMergeUpdates :: !(IORef [SomeMergeUpdate x])

eventEnvMergeInits :: !(IORef [SomeMergeInit x])

eventEnvClears :: !(IORef [SomeClear])

eventEnvRootClears :: !(IORef [SomeRootClear])

eventEnvCurrentHeight :: !(IORef Height)

eventEnvResetCoincidences :: !(IORef [SomeResetCoincidence x])

eventEnvDelayedMerges :: !(IORef (IntMap [EventM x ()]))

runEventM :: EventM x a -> IO a

asksEventEnv :: forall x a. HasSpiderTimeline x => (EventEnv x -> a) -> EventM x a

class MonadIO m => Defer a m where

Minimal complete definition

getDeferralQueue

Methods

getDeferralQueue :: m (IORef [a])

defer :: Defer a m => a -> m ()

class HasSpiderTimeline x => HasCurrentHeight x m | m -> x where

Minimal complete definition

getCurrentHeight, scheduleMerge

Methods

getCurrentHeight :: m Height

scheduleMerge :: Height -> EventM x () -> m ()

class HasSpiderTimeline x where

Minimal complete definition

spiderTimeline

Methods

spiderTimeline :: SpiderTimelineEnv x

Retrieve the current SpiderTimelineEnv

putCurrentHeight :: HasSpiderTimeline x => Height -> EventM x ()

scheduleClear :: Defer SomeClear m => IORef (Maybe a) -> m ()

scheduleRootClear :: Defer SomeRootClear m => IORef (DMap k Identity) -> m ()

hold :: (Patch p, Defer (SomeHoldInit x) m) => PatchTarget p -> Event x p -> m (Hold x p)

getHoldEventSubscription :: forall m p x. (Defer (SomeAssignment x) m, Defer (SomeHoldInit x) m, Defer SomeClear m, Patch p, HasCurrentHeight x m, Defer (SomeMergeInit x) m, Defer (SomeResetCoincidence x) m, HasSpiderTimeline x, m ~ EventM x) => Hold x p -> m (EventSubscription x)

type BehaviorEnv x = (Maybe (Weak (Invalidator x), IORef [SomeBehaviorSubscribed x]), IORef [SomeHoldInit x])

newtype BehaviorM x a

Constructors

BehaviorM

Fields

unBehaviorM :: ReaderT (BehaviorEnv x) IO a

data BehaviorSubscribed x a

Constructors

BehaviorSubscribedHold (Hold x p)

BehaviorSubscribedPull (PullSubscribed x a)

data SomeBehaviorSubscribed x

Constructors

SomeBehaviorSubscribed (BehaviorSubscribed x a)

data PullSubscribed x a

Constructors

PullSubscribed

Fields

pullSubscribedValue :: !a

pullSubscribedInvalidators :: !(IORef [Weak (Invalidator x)])

pullSubscribedOwnInvalidator :: !(Invalidator x)

pullSubscribedParents :: ![SomeBehaviorSubscribed x]

data Pull x a

Constructors

Pull

Fields

pullValue :: !(IORef (Maybe (PullSubscribed x a)))

pullCompute :: !(BehaviorM x a)

data Invalidator x

Constructors

InvalidatorPull (Pull x a)

InvalidatorSwitch (SwitchSubscribed x a)

data RootSubscribed x a

Constructors

GCompare k => RootSubscribed

Fields

rootSubscribedKey :: !(k a)

rootSubscribedCachedSubscribed :: !(IORef (DMap k (RootSubscribed x)))

rootSubscribedSubscribers :: !(WeakBag (Subscriber x a))

rootSubscribedOccurrence :: !(IO (Maybe a))

rootSubscribedUninit :: IO ()

rootSubscribedWeakSelf :: !(IORef (Weak (RootSubscribed x a)))

data Root x k

Constructors

Root

Fields

rootOccurrence :: !(IORef (DMap k Identity))

rootSubscribed :: !(IORef (DMap k (RootSubscribed x)))

rootInit :: !(forall a. k a -> RootTrigger x a -> IO (IO ()))

data SomeHoldInit x

Constructors

Patch p => SomeHoldInit !(Hold x p)

data SomeDynInit x

Constructors

Patch p => SomeDynInit !(Dyn x p)

data SomeMergeUpdate x

Constructors

SomeMergeUpdate

Fields

_someMergeUpdate_update :: !(EventM x [EventSubscription x])

_someMergeUpdate_invalidateHeight :: !(IO ())

_someMergeUpdate_recalculateHeight :: !(IO ())

newtype SomeMergeInit x

Constructors

SomeMergeInit

Fields

unSomeMergeInit :: EventM x ()

newtype EventM x a

Constructors

EventM

Fields

unEventM :: IO a

newtype ComputeM x a

Constructors

ComputeM (ReaderT (IORef [SomeHoldInit x], IORef [SomeDynInit x]) IO a)

runComputeM :: (Defer (SomeHoldInit x) m, Defer (SomeDynInit x) m) => ComputeM x a -> m a

newtype MergeSubscribedParent x a

Constructors

MergeSubscribedParent

Fields

unMergeSubscribedParent :: EventSubscription x

data MergeSubscribedParentWithMove x k a

Constructors

MergeSubscribedParentWithMove

Fields

_mergeSubscribedParentWithMove_subscription :: !(EventSubscription x)

_mergeSubscribedParentWithMove_key :: !(IORef (k a))

data HeightBag

Constructors

HeightBag

Fields

_heightBag_size :: !Int

_heightBag_contents :: !(IntMap Word)

heightBagEmpty :: HeightBag

heightBagSize :: HeightBag -> Int

heightBagFromList :: [Height] -> HeightBag

heightBagAdd :: Height -> HeightBag -> HeightBag

heightBagRemove :: Height -> HeightBag -> HeightBag

heightBagMax :: HeightBag -> Height

heightBagVerify :: HeightBag -> HeightBag

data FanSubscribedChildren x k a

Constructors

FanSubscribedChildren

Fields

_fanSubscribedChildren_list :: !(WeakBag (Subscriber x a))

_fanSubscribedChildren_self :: !(k a, FanSubscribed x k)

_fanSubscribedChildren_weakSelf :: !(IORef (Weak (k a, FanSubscribed x k)))

data FanSubscribed x k

Constructors

FanSubscribed

Fields

fanSubscribedCachedSubscribed :: !(IORef (Maybe (FanSubscribed x k)))

fanSubscribedOccurrence :: !(IORef (Maybe (DMap k Identity)))

fanSubscribedSubscribers :: !(IORef (DMap k (FanSubscribedChildren x k)))

fanSubscribedParent :: !(EventSubscription x)

data Fan x k

Constructors

Fan

Fields

fanParent :: !(Event x (DMap k Identity))

fanSubscribed :: !(IORef (Maybe (FanSubscribed x k)))

data SwitchSubscribed x a

Constructors

SwitchSubscribed

Fields

switchSubscribedCachedSubscribed :: !(IORef (Maybe (SwitchSubscribed x a)))

switchSubscribedOccurrence :: !(IORef (Maybe a))

switchSubscribedHeight :: !(IORef Height)

switchSubscribedSubscribers :: !(WeakBag (Subscriber x a))

switchSubscribedOwnInvalidator :: !(Invalidator x)

switchSubscribedOwnWeakInvalidator :: !(IORef (Weak (Invalidator x)))

switchSubscribedBehaviorParents :: !(IORef [SomeBehaviorSubscribed x])

switchSubscribedParent :: !(Behavior x (Event x a))

switchSubscribedCurrentParent :: !(IORef (EventSubscription x))

switchSubscribedWeakSelf :: !(IORef (Weak (SwitchSubscribed x a)))

data Switch x a

Constructors

Switch

Fields

switchParent :: !(Behavior x (Event x a))

switchSubscribed :: !(IORef (Maybe (SwitchSubscribed x a)))

data CoincidenceSubscribed x a

Constructors

CoincidenceSubscribed

Fields

coincidenceSubscribedCachedSubscribed :: !(IORef (Maybe (CoincidenceSubscribed x a)))

coincidenceSubscribedOccurrence :: !(IORef (Maybe a))

coincidenceSubscribedSubscribers :: !(WeakBag (Subscriber x a))

coincidenceSubscribedHeight :: !(IORef Height)

coincidenceSubscribedOuter :: (Subscriber x (Event x a))

coincidenceSubscribedOuterParent :: !(EventSubscription x)

coincidenceSubscribedInnerParent :: !(IORef (Maybe (EventSubscribed x)))

coincidenceSubscribedWeakSelf :: !(IORef (Weak (CoincidenceSubscribed x a)))

data Coincidence x a

Constructors

Coincidence

Fields

coincidenceParent :: !(Event x (Event x a))

coincidenceSubscribed :: !(IORef (Maybe (CoincidenceSubscribed x a)))

newInvalidatorSwitch :: SwitchSubscribed x a -> IO (Invalidator x)

newInvalidatorPull :: Pull x a -> IO (Invalidator x)

newtype Dyn x p

Constructors

Dyn

Fields

unDyn :: IORef (Either (BehaviorM x (PatchTarget p), Event x p) (Hold x p))

newMapDyn :: HasSpiderTimeline x => (a -> b) -> Dynamic x (Identity a) -> Dynamic x (Identity b)

zipDynWith :: HasSpiderTimeline x => (a -> b -> c) -> Dynamic x (Identity a) -> Dynamic x (Identity b) -> Dynamic x (Identity c)

buildDynamic :: (Defer (SomeDynInit x) m, Patch p) => BehaviorM x (PatchTarget p) -> Event x p -> m (Dyn x p)

unsafeBuildDynamic :: BehaviorM x (PatchTarget p) -> Event x p -> Dyn x p

type ResultM = EventM

unsafeNewIORef :: a -> b -> IORef b

push :: HasSpiderTimeline x => (a -> ComputeM x (Maybe b)) -> Event x a -> Event x b

pull :: BehaviorM x a -> Behavior x a

switch :: HasSpiderTimeline x => Behavior x (Event x a) -> Event x a

coincidence :: HasSpiderTimeline x => Event x (Event x a) -> Event x a

run :: forall x b. HasSpiderTimeline x => [DSum (RootTrigger x) Identity] -> ResultM x b -> SpiderHost x b

scheduleMerge' :: HasSpiderTimeline x => Height -> IORef Height -> EventM x () -> EventM x ()

data SomeClear

Constructors

SomeClear (IORef (Maybe a))

data SomeRootClear

Constructors

SomeRootClear (IORef (DMap k Identity))

data SomeAssignment x

Constructors

SomeAssignment !(IORef a) !(IORef [Weak (Invalidator x)]) a

debugFinalize :: Bool

mkWeakPtrWithDebug :: a -> String -> IO (Weak a)

type WeakList a = [Weak a]

withIncreasedDepth :: m a -> m a

type CanTrace x m = (HasSpiderTimeline x, MonadIO m)

tracePropagate :: CanTrace x m => proxy x -> String -> m ()

traceInvalidate :: String -> IO ()

traceWhen :: CanTrace x m => proxy x -> Bool -> String -> m ()

traceMWhen :: CanTrace x m => proxy x -> Bool -> m String -> m ()

whoCreatedIORef :: IORef a -> IO [String]

propagateSubscriberHold :: forall x p. (HasSpiderTimeline x, Patch p) => Hold x p -> p -> EventM x ()

data SomeResetCoincidence x

Constructors

SomeResetCoincidence !(EventSubscription x) !(Maybe (CoincidenceSubscribed x a))

runBehaviorM :: BehaviorM x a -> Maybe (Weak (Invalidator x), IORef [SomeBehaviorSubscribed x]) -> IORef [SomeHoldInit x] -> IO a

askInvalidator :: BehaviorM x (Maybe (Weak (Invalidator x)))

askParentsRef :: BehaviorM x (Maybe (IORef [SomeBehaviorSubscribed x]))

askBehaviorHoldInits :: BehaviorM x (IORef [SomeHoldInit x])

getDynHold :: (Defer (SomeHoldInit x) m, Patch p) => Dyn x p -> m (Hold x p)

zeroRef :: IORef Height

type CanSubscribe x m = (HasSpiderTimeline x, HasCurrentHeight x m, Defer (SomeHoldInit x) m, Defer (SomeMergeInit x) m, Defer SomeClear m, Defer (SomeAssignment x) m, Defer (SomeResetCoincidence x) m, m ~ EventM x)

getRootSubscribed :: GCompare k => k a -> Root x k -> Subscriber x a -> IO (WeakBagTicket, RootSubscribed x a, Maybe a)

cleanupRootSubscribed :: RootSubscribed x a -> IO ()

subscribeRootSubscribed :: RootSubscribed x a -> Subscriber x a -> IO WeakBagTicket

getFanSubscribed :: (Defer (SomeAssignment x) m, Defer (SomeHoldInit x) m, Defer (SomeMergeInit x) m, Defer SomeClear m, HasCurrentHeight x m, Defer (SomeResetCoincidence x) m, HasSpiderTimeline x, m ~ EventM x, GCompare k) => k a -> Fan x k -> Subscriber x a -> m (WeakBagTicket, FanSubscribed x k, Maybe a)

cleanupFanSubscribed :: GCompare k => (k a, FanSubscribed x k) -> IO ()

subscribeFanSubscribed :: GCompare k => k a -> FanSubscribed x k -> Subscriber x a -> IO WeakBagTicket

getSwitchSubscribed :: (Defer (SomeHoldInit x) m, Defer (SomeAssignment x) m, Defer (SomeMergeInit x) m, Defer SomeClear m, Defer (SomeResetCoincidence x) m, HasCurrentHeight x m, HasSpiderTimeline x, m ~ EventM x) => Switch x a -> Subscriber x a -> m (WeakBagTicket, SwitchSubscribed x a, Maybe a)

cleanupSwitchSubscribed :: SwitchSubscribed x a -> IO ()

subscribeSwitchSubscribed :: SwitchSubscribed x a -> Subscriber x a -> IO WeakBagTicket

getCoincidenceSubscribed :: forall x a m. (Defer (SomeAssignment x) m, Defer (SomeHoldInit x) m, Defer (SomeMergeInit x) m, Defer SomeClear m, HasCurrentHeight x m, Defer (SomeResetCoincidence x) m, HasSpiderTimeline x, m ~ EventM x) => Coincidence x a -> Subscriber x a -> m (WeakBagTicket, CoincidenceSubscribed x a, Maybe a)

cleanupCoincidenceSubscribed :: CoincidenceSubscribed x a -> IO ()

subscribeCoincidenceSubscribed :: CoincidenceSubscribed x a -> Subscriber x a -> IO WeakBagTicket

merge :: forall k x. (HasSpiderTimeline x, GCompare k) => Dynamic x (PatchDMap k (Event x)) -> Event x (DMap k Identity)

mergeWithMove :: forall k x. (HasSpiderTimeline x, GCompare k) => Dynamic x (PatchDMapWithMove k (Event x)) -> Event x (DMap k Identity)

mergeCheap :: forall k x. (HasSpiderTimeline x, GCompare k) => Dynamic x (PatchDMap k (Event x)) -> Event x (DMap k Identity)

mergeCheapWithMove :: forall k x. (HasSpiderTimeline x, GCompare k) => Dynamic x (PatchDMapWithMove k (Event x)) -> Event x (DMap k Identity)

type MergeUpdateFunc k x p s = (forall a. EventM x (k a) -> Subscriber x a) -> IORef HeightBag -> DMap k s -> p -> EventM x ([EventSubscription x], DMap k s)

type MergeInitFunc k x s = DMap k (Event x) -> (forall a. EventM x (k a) -> Subscriber x a) -> EventM x (DMap k Identity, [Height], DMap k s)

type MergeDestroyFunc k s = DMap k s -> IO ()

mergeCheap' :: forall k x p s. (HasSpiderTimeline x, GCompare k, PatchTarget p ~ DMap k (Event x)) => MergeInitFunc k x s -> MergeUpdateFunc k x p s -> MergeDestroyFunc k s -> (DMap k s -> Int) -> Dynamic x p -> Event x (DMap k Identity)

newtype EventSelector x k

Constructors

EventSelector

Fields

select :: forall a. k a -> Event x a

fan :: (HasSpiderTimeline x, GCompare k) => Event x (DMap k Identity) -> EventSelector x k

runHoldInits :: HasSpiderTimeline x => IORef [SomeHoldInit x] -> IORef [SomeDynInit x] -> IORef [SomeMergeInit x] -> EventM x ()

initHold :: HasSpiderTimeline x => SomeHoldInit x -> EventM x ()

initDyn :: HasSpiderTimeline x => SomeDynInit x -> EventM x ()

newEventEnv :: IO (EventEnv x)

clearEventEnv :: EventEnv x -> IO ()

runFrame :: forall x a. HasSpiderTimeline x => EventM x a -> SpiderHost x a

Run an event action outside of a frame

newtype Height

Constructors

Height

Fields

unHeight :: Int

zeroHeight :: Height

invalidHeight :: Height

succHeight :: Height -> Height

invalidateCoincidenceHeight :: CoincidenceSubscribed x a -> IO ()

updateSwitchHeight :: Height -> SwitchSubscribed x a -> IO ()

recalculateCoincidenceHeight :: CoincidenceSubscribed x a -> IO ()

calculateSwitchHeight :: SwitchSubscribed x a -> IO Height

calculateCoincidenceHeight :: CoincidenceSubscribed x a -> IO Height

data SomeSwitchSubscribed x

Constructors

SomeSwitchSubscribed (SwitchSubscribed x a)

invalidate :: IORef [SomeSwitchSubscribed x] -> WeakList (Invalidator x) -> IO (WeakList (Invalidator x))

unsafeNewSpiderTimelineEnv :: forall x. IO (SpiderTimelineEnv x)

newSpiderTimeline :: IO (Some SpiderTimelineEnv)

Create a new SpiderTimelineEnv

data LocalSpiderTimeline x s

localSpiderTimeline :: Proxy s -> SpiderTimelineEnv x -> SpiderTimelineEnv (LocalSpiderTimeline x s)

withSpiderTimeline :: (forall x. HasSpiderTimeline x => SpiderTimelineEnv x -> IO r) -> IO r

Pass a new timeline to the given function.

newtype SpiderPullM x a

Constructors

SpiderPullM (BehaviorM x a)

newtype SpiderPushM x a

Constructors

SpiderPushM (ComputeM x a)

data RootTrigger x a

Constructors

GCompare k => RootTrigger (WeakBag (Subscriber x a), IORef (DMap k Identity), k a)

data SpiderEventHandle x a

Constructors

SpiderEventHandle

Fields

spiderEventHandleSubscription :: EventSubscription x

spiderEventHandleValue :: IORef (Maybe a)

newtype SpiderHost x a

The monad for actions that manipulate a Spider timeline identified by x

Constructors

SpiderHost

Fields

unSpiderHost :: ReaderT (SpiderTimelineEnv x) IO a

runSpiderHost :: SpiderHost Global a -> IO a

Run an action affecting the global Spider timeline; this will be guarded by a mutex for that timeline

runSpiderHostForTimeline :: SpiderHost x a -> SpiderTimelineEnv x -> IO a

Run an action affecting a given Spider timeline; this will be guarded by a mutex for that timeline

newtype SpiderHostFrame x a

Constructors

SpiderHostFrame

Fields

runSpiderHostFrame :: EventM x a

newEventWithTriggerIO :: (RootTrigger x a -> IO (IO ())) -> IO (Event x a)

newFanEventWithTriggerIO :: GCompare k => (forall a. k a -> RootTrigger x a -> IO (IO ())) -> IO (EventSelector x k)

newtype ReadPhase x a

Constructors

ReadPhase (ResultM x a)

data SpiderTimeline x

Designates the default, global Spider timeline

type Spider = SpiderTimeline Global

The default, global Spider environment

newJoinDyn :: HasSpiderTimeline x => Dynamic x (Identity (Dynamic x (Identity a))) -> Dyn x (Identity a)

holdSpiderEventM :: HasSpiderTimeline x => a -> Event (SpiderTimeline x) a -> EventM x (Behavior (SpiderTimeline x) a)

holdDynSpiderEventM :: HasSpiderTimeline x => a -> Event (SpiderTimeline x) a -> EventM x (Dynamic (SpiderTimeline x) a)

holdIncrementalSpiderEventM :: (HasSpiderTimeline x, Patch p) => PatchTarget p -> Event (SpiderTimeline x) p -> EventM x (Incremental (SpiderTimeline x) p)

buildDynamicSpiderEventM :: HasSpiderTimeline x => SpiderPullM x a -> Event (SpiderTimeline x) a -> EventM x (Dynamic (SpiderTimeline x) a)

type SpiderEnv = SpiderTimeline

Deprecated: Use SpiderTimelineEnv instead

SpiderEnv is the old name for SpiderTimeline